+++ applications/dataviews-editor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @globalfishingwatchapp/dataviews-editor
 
+## 1.1.20
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@4.17.3
+
 ## 1.1.19
 
 ### Patch Changes